python/requirements.txt
beautifulsoup4
imgkit
qrcode
Requests
selenium
twocaptcha
syncer
pyppeteer
sqlalchemy
mysql-connector-python
python-dotenv
cryptography